move fetch to server ts update order status
