+++ frontend/src/routes/order/updateOrderStatus/+page.svelte
@@ -1,10 +1,40 @@
 <script lang="ts">
 	import { goto, invalidateAll } from '$app/navigation';
-	import { backendPath } from '$lib/components/PUA/env';
 	import OrderStatusCard from '$lib/components/PUA/orderStatusCard.svelte';
 	import { onMount } from 'svelte';
 	import { page } from '$app/stores';
+	import type { PageData } from './$types';
+	import { backendPath } from '$lib/components/PUA/env';
 
+	export let data: PageData;
+
+	async function postAndChangeStatus(index: number) {
+		let orderRespList = data.orderRespList;
+		await fetch(
+			backendPath +
+				`/seller/update-order-status/customer/` +
+				orderRespList[index].user_id +
+				`/cart/` +
+				orderRespList[index].cart_id +
+				`/store/` +
+				orderRespList[index].store_id,
+			{
+				method: 'PUT'
+			}
+		);
+		await invalidateAll();
+
+		await refresh();
+		return;
+	}
+	let statusCardContent: {
+		time: string;
+		price: string;
+		src: string;
+		text: string;
+		user: string;
+		status: number;
+	}[] = [];
 	let Deliver: string =
 		'M48 0C21.5 0 0 21.5 0 48V368c0 26.5 21.5 48 48 48H64c0 53 43 96 96 96s96-43 96-96H384c0 53 43 96 96 96s96-43 96-96h32c17.7 0 32-14.3 32-32s-14.3-32-32-32V288 256 237.3c0-17-6.7-33.3-18.7-45.3L512 114.7c-12-12-28.3-18.7-45.3-18.7H416V48c0-26.5-21.5-48-48-48H48zM416 160h50.7L544 237.3V256H416V160zM112 416a48 48 0 1 1 96 0 48 48 0 1 1 -96 0zm368-48a48 48 0 1 1 0 96 48 48 0 1 1 0-96z';
 	let Accept: string =
@@ -19,85 +49,27 @@
 		{ icon: Making, text: 'Deliver' },
 		{ icon: Deliver, text: 'Arrival' }
 	];
-
-	type orderRespTye = {
-		store_id: number;
-		cart_id: number;
-		order_date: string;
-		total_price: number;
-		user_id: number;
-		status: number;
-		user_name: string;
-	};
-	let orderRespList: orderRespTye[] = [];
-	let statusCardContent: {
-		time: string;
-		price: string;
-		src: string;
-		text: string;
-		user: string;
-		status: number;
-	}[] = [];
-
-	async function postAndChangeStatus(index: number) {
-		const Resp = await fetch(
-			backendPath +
-				`/seller/update-order-status/customer/` +
-				orderRespList[index].user_id +
-				`/cart/` +
-				orderRespList[index].cart_id +
-				`/store/` +
-				orderRespList[index].store_id,
-			{
-				method: 'PUT'
-			}
-		);
-
-		let orderResp = await Resp.json();
-
-		if (orderResp.length != 0 && orderResp.status < 5) {
-			statusCardContent[index] = {
-				time: orderResp.order_date,
-				price: orderResp.total_price.toString(),
-				src: icon_array[orderResp.status - 1].icon,
-				text: icon_array[orderResp.status - 1].text,
-				user: orderResp.user_name,
-				status: orderResp.status
-			};
-		} else {
-			statusCardContent = statusCardContent.filter((_, i) => i !== index);
-		}
-
-		invalidateAll();
-
-		return;
-	}
-	async function getOrderRespList() {
-		const Resp = await fetch(backendPath + '/seller/store/' + '1' + '/orders');
-		if (Resp.status == 200) {
-			orderRespList = await Resp.json();
-
-			console.log(orderRespList);
-
-			if (orderRespList) {
-				for (let i = 0; i < orderRespList.length; i++) {
-					statusCardContent = [
-						...statusCardContent,
-						{
-							time: orderRespList[i].order_date,
-							price: orderRespList[i].total_price.toString(),
-							src: icon_array[orderRespList[i].status - 1].icon,
-							text: icon_array[orderRespList[i].status - 1].text,
-							user: orderRespList[i].user_name,
-							status: orderRespList[i].status
-						}
-					];
-				}
+	async function refresh() {
+		statusCardContent = [];
+		let orderRespList = data.orderRespList;
+		if (orderRespList) {
+			for (let i = 0; i < orderRespList.length; i++) {
+				statusCardContent = [
+					...statusCardContent,
+					{
+						time: orderRespList[i].order_date,
+						price: orderRespList[i].total_price.toString(),
+						src: icon_array[orderRespList[i].status - 1].icon,
+						text: icon_array[orderRespList[i].status - 1].text,
+						user: orderRespList[i].user_name,
+						status: orderRespList[i].status
+					}
+				];
 			}
 		}
 	}
 	onMount(async () => {
-		getOrderRespList();
+		await refresh();
 	});
 </script>
 
@@ -108,8 +80,8 @@
 			on:gotoPage={(event) => {
 				goto($page.route.id + '/detail/' + event.detail.storeId + '/' + event.detail.cartId);
 			}}
-			storeId={orderRespList[index].store_id}
-			cartId={orderRespList[index].cart_id}
+			storeId={data.orderRespList[index].store_id}
+			cartId={data.orderRespList[index].cart_id}
 			statusCardContent={sub}
 		></OrderStatusCard>
 	{/each}
